Add image width, height, and loading attributes
The loading attribute will help with lazy loading in Chromium browsers 76+ and hopefully other browsers in the future.

The dimensional attributes will help in the future by allowing the browser to calculate an early aspect ratio for the image before downloading, allowing some space to be reserved and hopefully preventing content shifts from happening.